Package renaming, billboard selection added
