--- mobile/app/src/main/java/com/vcelicky/smog/models/Photo.java
@@ -1,56 +1,0 @@
-package com.vcelicky.smog.models;
-
-import java.io.Serializable;
-
-/**
- * Created by jerry on 24. 11. 2014.
- */
-public class Photo implements Serializable {
-
-    private byte[] imageByteArray;
-    private double latitude;
-    private double longitude;
-    private String comment;
-    private String model;
-
-    public byte[] getImageByteArray() {
-        return imageByteArray;
-    }
-
-    public void setImageByteArray(byte[] imageByteArray) {
-        this.imageByteArray = imageByteArray;
-    }
-
-    public double getLatitude() {
-        return latitude;
-    }
-
-    public void setLatitude(double latitude) {
-        this.latitude = latitude;
-    }
-
-    public double getLongitude() {
-        return longitude;
-    }
-
-    public void setLongitude(double longitude) {
-        this.longitude = longitude;
-    }
-
-    public String getComment() {
-        return comment;
-    }
-
-    public void setComment(String comment) {
-        this.comment = comment;
-    }
-
-    public String getModel() {
-        return model;
-    }
-
-    public void setModel(String model) {
-        this.model = model;
-    }
-
-}
