+++ mobile/app/src/main/java/sk/fiit/adhunter/models/Photo.java
@@ -1,0 +1,65 @@
+package sk.fiit.adhunter.models;
+
+import java.io.Serializable;
+
+/**
+ * Created by jerry on 24. 11. 2014.
+ */
+public class Photo implements Serializable {
+
+    private byte[] imageByteArray;
+    private double latitude;
+    private double longitude;
+    private String comment;
+    private String billboardType;
+    private String owner;
+
+    public byte[] getImageByteArray() {
+        return imageByteArray;
+    }
+
+    public void setImageByteArray(byte[] imageByteArray) {
+        this.imageByteArray = imageByteArray;
+    }
+
+    public double getLatitude() {
+        return latitude;
+    }
+
+    public void setLatitude(double latitude) {
+        this.latitude = latitude;
+    }
+
+    public double getLongitude() {
+        return longitude;
+    }
+
+    public void setLongitude(double longitude) {
+        this.longitude = longitude;
+    }
+
+    public String getComment() {
+        return comment;
+    }
+
+    public void setComment(String comment) {
+        this.comment = comment;
+    }
+
+    public String getBillboardType() {
+        return billboardType;
+    }
+
+    public void setBillboardType(String billboardType) {
+        this.billboardType = billboardType;
+    }
+
+    public String getOwner() {
+        return owner;
+    }
+
+    public void setOwner(String owner) {
+        this.owner = owner;
+    }
+
+}
